Add core directory and separate out core stuff from parse
